refactor: move KintoneAllRecordsError as a seperated file
